00-namespaces/namespaces-mcp/level1/tools/requirements.txt
# MCP Level 1 Tools Requirements

# Core dependencies
pyyaml>=6.0.1
jsonschema>=4.20.0

# CLI tools
click>=8.1.7
rich>=13.7.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Code quality
pylint>=3.0.3
black>=23.12.1
mypy>=1.8.0